+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Fixed
-- Ressources Getter - Set order to undefined if the query is empty.
+- Primary Composite Key - Ressource remover handle primary composite key. 
 
 ## RELEASE 1.1.9 - 2017-06-28
 ### Fixed